fix fetcher's test comments
